split song table into ui and row components
